chore: add new extensions for mapping the HearingDevicePairedEar to String
