+++ Source/Features/Extensions/UIContentSizeCategory+values.swift
@@ -40,6 +40,10 @@ extension UIContentSizeCategory {
             return "Unknown"
         }
     }
+
+    var isDefault: Bool {
+        return self == .medium
+    }
 }
 
 #endif
